+++ CHANGELOG.md
@@ -25,6 +25,9 @@
 * Quote names containing colons in generated .d.ts.
   [#4488](https://github.com/wasm-bindgen/wasm-bindgen/pull/4488)
 
+* Fixes TryFromJsValue for structs JsValue stack corruption on failure.
+  [#4786](https://github.com/wasm-bindgen/wasm-bindgen/pull/4786)
+
 ### Removed
 
 ## [0.2.105](https://github.com/wasm-bindgen/wasm-bindgen/compare/0.2.104...0.2.105)